+++ UnleashClient/strategies/ApplicationHostname.py
@@ -1,12 +1,12 @@
 import platform
-from UnleashClient.strategies.Strategies import StrategyV2
+from UnleashClient.strategies.Strategies import Strategy
 
 
-class ApplicationHostname(StrategyV2):
+class ApplicationHostname(Strategy):
     def load_provisioning(self) -> list:
         return [x.strip() for x in self.parameters["hostNames"].split(',')]
 
-    def apply_strategy(self, context: dict = None) -> bool:
+    def apply(self, context: dict = None) -> bool:
         """
         Returns true if userId is a member of id list.