+++ UnleashClient/strategies/UserWithId.py
@@ -1,11 +1,11 @@
-from UnleashClient.strategies.Strategies import StrategyV2
+from UnleashClient.strategies.Strategies import Strategy
 
 
-class UserWithId(StrategyV2):
+class UserWithId(Strategy):
     def load_provisioning(self) -> list:
         return [x.strip() for x in self.parameters["userIds"].split(',')]
 
-    def apply_strategy(self, context: dict = None) -> bool:
+    def apply(self, context: dict = None) -> bool:
         """
         Returns true if userId is a member of id list.