+++ tests/unit_tests/strategies/test_flexiblerollout.py
@@ -57,7 +57,7 @@ def test_flexiblerollout_satisfiesconstraints(strategy):
         'environment': 'prod'
     }
 
-    assert strategy(context)
+    assert strategy.execute(context)
 
 
 def test_flexiblerollout_doesntsatisfiesconstraints(strategy):
@@ -66,30 +66,30 @@ def test_flexiblerollout_doesntsatisfiesconstraints(strategy):
         'appName': 'qualityhamster',
         'environment': 'prod'
     }
-    assert not strategy(context)
+    assert not strategy.execute(context)
 
 
 def test_flexiblerollout_userid(strategy):
     base_context = dict(appName='test', environment='prod')
     base_context['userId'] = "122"
-    assert strategy(base_context)
+    assert strategy.execute(base_context)
     base_context['userId'] = "155"
-    assert not strategy(base_context)
+    assert not strategy.execute(base_context)
 
 
 def test_flexiblerollout_sessionid(strategy):
     BASE_FLEXIBLE_ROLLOUT_DICT['parameters']['stickiness'] = 'sessionId'
     base_context = dict(appName='test', environment='prod', userId="9")
     base_context['sessionId'] = "122"
-    assert strategy(base_context)
+    assert strategy.execute(base_context)
     base_context['sessionId'] = "155"
-    assert not strategy(base_context)
+    assert not strategy.execute(base_context)
 
 
 def test_flexiblerollout_random(strategy):
     BASE_FLEXIBLE_ROLLOUT_DICT['parameters']['stickiness'] = 'random'
     base_context = dict(appName='test', environment='prod', userId="1")
-    assert strategy(base_context) in [True, False]
+    assert strategy.execute(base_context) in [True, False]
 
 
 def test_flexiblerollout_default():
@@ -97,8 +97,8 @@ def test_flexiblerollout_default():
     BASE_FLEXIBLE_ROLLOUT_DICT['constraints'] = [x for x in BASE_FLEXIBLE_ROLLOUT_DICT['constraints'] if x['contextName'] != 'userId']
     strategy = FlexibleRollout(BASE_FLEXIBLE_ROLLOUT_DICT['constraints'], BASE_FLEXIBLE_ROLLOUT_DICT['parameters'])
     base_context = dict(appName='test', environment='prod', userId="122", sessionId="155")
-    assert strategy(base_context)
+    assert strategy.execute(base_context)
     base_context = dict(appName='test', environment='prod', sessionId="122")
-    assert strategy(base_context)
+    assert strategy.execute(base_context)
     base_context = dict(appName='test', environment='prod')
-    assert strategy(base_context) in [True, False]
+    assert strategy.execute(base_context) in [True, False]