+++ tests/unit_tests/strategies/test_gradualrolloutwithsessionid.py
@@ -13,4 +13,4 @@ def strategy():
 
 
 def test_userwithid(strategy):
-    strategy(context=generate_context())
+    strategy.execute(context=generate_context())